Add script to fetch releases and compute hashes
This commit adds a script that contains several functions that can all
be used individually, or chained together to fetch the kubernetes
releases and then traverse the list, fetch each tarball and compute a
Nix hash for it. The resulting file is added to Git, because it's used
to generate lots of "fetchurl" calls.

The lib.fish script contains a "run" function which does the function
call chaining for you.