Committing this for posterity
This commit implements one approach I took for this package where I just
download the release tags from GitHub and use that to generate the Nix
Flake inputs.

The problem is that downloading older sources means I also need to apply
the older Nixpkgs derivation, and that's way too much work.

I won't use this commit for the Flake and do something else instead, but
I didn't want to lose this code.